# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Lending Platform API
  description: |
    REST API for a simplified lending platform backend.
    See [API Design doc](03-api-and-schema.md) for full context.
  version: 0.1.0
  contact:
    name: Engineering Lead
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
...
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Client-generated UUID ensuring at-most-once execution.
      schema:
        type: string
        format: uuid
        example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
    XUserId:
      name: X-User-Id
      in: header
      required: false
      description: Stub user identifier for audit trail.
      schema:
        type: string
        example: "admin-001"
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 0
    SizeParam:
      name: size
      in: query
      required: false
      schema:
        type: integer
        default: 20
    SortParam:
      name: sort
      in: query
      required: false
      schema:
        type: string
        example: "createdAt,desc"
  schemas:
...
    Money:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: string
          description: Decimal value with up to 4 decimal places.
          example: "10000.0000"
        currency:
          type: string
          minLength: 3
          maxLength: 3
          description: ISO 4217 currency code.
          example: "USD"
...
    LoanAccountResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        applicationId:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        principal:
          $ref: '#/components/schemas/Money'
        interestRatePerAnnum:
          type: string
        interestAccrualMethod:
          type: string
          enum: [PRE_COMPUTED, DAILY_ACCRUAL]
        tenureMonths:
          type: integer
        repaymentFrequency:
          type: string
          enum: [MONTHLY, BI_WEEKLY, WEEKLY]
        gracePeriodDays:
          type: integer
        originationFeeChargeMethod:
          type: string
          enum: [DEDUCTED_FROM_DISBURSEMENT, ADDED_TO_BALANCE]
        allowEarlyRepayment:
          type: boolean
        overpaymentPolicy:
          type: string
          enum: [REJECT, APPLY_TO_NEXT_INSTALLMENT, HOLD_AS_CREDIT]
        originationFee:
          $ref: '#/components/schemas/Money'
        outstandingBalance:
          $ref: '#/components/schemas/Money'
        creditBalance:
          $ref: '#/components/schemas/Money'
          description: Surplus held under HOLD_AS_CREDIT overpayment policy.
        state:
          type: string
          enum: [PENDING_DISBURSEMENT, ACTIVE, OVERDUE, DEFAULTED, WRITTEN_OFF, CLOSED]
        disbursedAt:
          type: string
          format: date-time
          nullable: true
        closedAt:
          type: string
          format: date-time
          nullable: true
        nextInstallmentDueDate:
          type: string
          format: date
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        version:
          type: integer
          format: int64
    InstallmentResponse:
      type: object
      properties:
        installmentNumber:
          type: integer
        dueDate:
          type: string
          format: date
        principalDue:
          $ref: '#/components/schemas/Money'
        interestDue:
          $ref: '#/components/schemas/Money'
        totalDue:
          $ref: '#/components/schemas/Money'
        principalPaid:
          $ref: '#/components/schemas/Money'
        interestPaid:
          $ref: '#/components/schemas/Money'
        totalPaid:
          $ref: '#/components/schemas/Money'
        status:
          type: string
          enum: [PENDING, PARTIALLY_PAID, PAID, OVERDUE]
    ScheduleResponse:
      type: object
      properties:
        loanAccountId:
          type: string
          format: uuid
        installments:
          type: array
          items:
            $ref: '#/components/schemas/InstallmentResponse'

    # ---- Repayments ----
    RecordRepaymentRequest:
      type: object
      required: [amount]
      properties:
        amount:
          $ref: '#/components/schemas/Money'
    RepaymentAllocation:
      type: object
      properties:
        installmentNumber:
          type: integer
        principalAllocated:
          $ref: '#/components/schemas/Money'
        interestAllocated:
          $ref: '#/components/schemas/Money'
    RepaymentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        loanAccountId:
          type: string
          format: uuid
        amount:
          $ref: '#/components/schemas/Money'
        receivedAt:
          type: string
          format: date-time
        allocations:
          type: array
          items:
            $ref: '#/components/schemas/RepaymentAllocation'
        remainingBalance:
          $ref: '#/components/schemas/Money'
        loanState:
          type: string
          enum: [ACTIVE, OVERDUE, CLOSED]
        createdAt:
          type: string
          format: date-time

    # ---- Ledger ----
    LedgerEntryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        loanAccountId:
          type: string
          format: uuid
        entryType:
          type: string
          enum: [DISBURSEMENT, REPAYMENT_PRINCIPAL, REPAYMENT_INTEREST, FEE_CHARGE, FEE_PAYMENT, WRITE_OFF]
        amount:
          $ref: '#/components/schemas/Money'
        direction:
          type: string
          enum: [DEBIT, CREDIT]
        description:
          type: string
        referenceId:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time

    # ---- Errors ----
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        code:
          type: string
          description: Machine-readable error code.
          example: "VALIDATION_ERROR"
        message:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
        path:
          type: string
        traceId:
          type: string

    # ---- Pagination wrapper ----
    PageMetadata:
      type: object
      properties:
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
  responses:
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest:
      description: Validation error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict — duplicate key, optimistic lock failure, or invalid state transition.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnprocessableEntity:
      description: Business rule violation (e.g., eligibility failure).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /loans:
    get:
      tags: [Loans]
      summary: List loan accounts
      operationId: listLoans
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
        - name: state
          in: query
          required: false
          schema:
            type: string
            enum: [PENDING_DISBURSEMENT, ACTIVE, OVERDUE, DEFAULTED, WRITTEN_OFF, CLOSED]
        - name: customerId
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Paginated list of loan accounts.
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/LoanAccountResponse'
                  page:
                    $ref: '#/components/schemas/PageMetadata'
  /loans/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Loans]
      summary: Get loan account details
      operationId: getLoan
      responses:
        '200':
          description: Loan account details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanAccountResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /loans/{id}/disburse:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Loans]
      summary: Disburse the loan
      operationId: disburseLoan
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/XUserId'
      responses:
        '200':
          description: Loan disbursed and active.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanAccountResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /loans/{id}/schedule:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Loans]
      summary: Get amortisation schedule
      operationId: getLoanSchedule
      responses:
        '200':
          description: Full amortisation schedule.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /loans/{id}/ledger:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Loans]
      summary: Get ledger entries for a loan
      operationId: getLoanLedger
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Paginated ledger entries.
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/LedgerEntryResponse'
                  page:
                    $ref: '#/components/schemas/PageMetadata'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---- Repayments ----
  /loans/{loanId}/repayments:
    parameters:
      - name: loanId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Repayments]
      summary: Record a repayment
      operationId: recordRepayment
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/XUserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordRepaymentRequest'
      responses:
        '201':
          description: Repayment recorded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepaymentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
    get:
      tags: [Repayments]
      summary: List repayments for a loan
      operationId: listRepayments
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Paginated list of repayments.
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/RepaymentResponse'
                  page:
                    $ref: '#/components/schemas/PageMetadata'
        '404':
          $ref: '#/components/responses/NotFound'